Split editor instance into input logic and instance
Base for further work on the split and input cache expected caching
